services/ch-mv-optimizer/requirements.txt
clickhouse-driver==0.2.6
pyyaml==6.0.1
prometheus-client==0.19.0
schedule==1.2.0
sqlparse==0.4.4
